Added more robust new sector routine. New sectors are no longer automatically loaded from the newsector.xml schema. Only sectors flagged for addition in the Remap Schema will be pulled from the newsectors.xml. Also added console progress updates on the remapping process.
