+++ schemas/readtext_descriptions.csv
@@ -1,0 +1,1 @@
+id,text0,Flaming Dawn0,Evaporated Skies0,Purple Fog0,The Twins0,Dual Eclipse0,Three Moons0,City in Space0,Thor's Might0,Accretion Disc0,Rhonkar's Clouds0,Guiding Star0,Prophecy Fulfilled1030101,Kho's Vengeance1030102,Brennan's Memorial1030117,Duke's Retreat1030118,Fire of Redemption1030121,"{7,1020000} {12,123}"1030122,"{17,10821} {12,110}"1030201,Arcturus Prime1030202,Eye of Colossus1030217,Consecrated Fire1030218,Heaven's Assertion1030222,"{17,10531} {12,109}"1030223,"{17,10431} {12,108}"1030301,Starkiller's Clan1030302,Pilgrim's Relay1030303,Kuiper Belt1030304,Pluto1030305,Neptune1030306,Uranus1030307,Jupiter1030308,Saturn1030312,Scion Fire1030313,Spring of Belief1030314,Perpetual Sin1030315,Harmony1030317,Grand Exchange1030318,Tears of Greed1030319,Elena's Fortune1030320,"{17,10131} {12,100}"1030322,"{17,10231} {12,106}"1030323,"{17,10321} {12,107}"1030401,Starkiller's Hideout1030402,Family Zein1030403,Family Ryk1030404,Oort Cloud1030405,The Moon1030406,Earth1030407,Mars1030408,Family Njy1030409,Njy's Deception1030412,Emperor Mines1030413,Paranid Prime1030414,Priest Rings1030415,Elysium of Light1030416,Unknown Sector Gamma1030417,Merchant Haven1030418,Xenon Sector 3471030419,Priest's Pity1030420,"{17,10261} {12,101}"1030421,"{17,10331} {12,102}"1030422,"{17,10191} {12,105}"1030501,Smuggler's Path1030502,Y Gamma1030503,Rhonkar's Trial1030504,Family Dispute1030505,Mercury1030506,Venus1030507,Asteroid Belt1030508,Thyn's Abyss1030509,Patriarch's Conclusion1030510,Thyn's Excavation1030511,Unseen Domain1030512,Emperor's Pride1030513,Unholy Descent1030514,Ceo's Buckzoid1030515,Teladi Gain1030516,Third Fortune1030517,Lost Profit1030518,Unknown Sector Beta1030519,Emperor's Ridge1030520,Imperial Frontier1030521,"{17,10451} {12,103}"1030522,"{17,10501} {12,104}"1030602,Wretched Skies1030603,The Shallows1030604,Overmasters1030605,Icarus1030606,Heretic's End1030607,Farnham's Legend1030608,Bala Gi's Joy1030609,Olmancketslat's Treaty1030610,Contorted Dominion1030611,Xenon Sector 4721030612,Nomad's Gate1030613,Unknown Sector Kappa1030614,Profit Share1030615,Seizewell1030617,Unholy Gift1030618,Duke's Citadel1030619,Pontifex' Realm1030620,Emperor's Glory1030703,Light Water1030704,Menelaus' Paradise1030705,Rolk's Legacy1030706,Great Trench1030707,Cloudbase North West1030708,Herron's Nebula1030709,The Hole1030710,Split Fire1030711,Sanctuary of Darkness1030712,Xenon Sector 1011030713,Profit Guard1030714,Spaceweed Drift1030715,Greater Profit1030718,Three Worlds1030719,Power Circle1030720,Antigone Memorial1030803,Shore of Infinity1030804,Bluish Snout1030805,Hila's Joy1030806,Ocean of Fantasy1030807,Ringo Moon1030808,Argon Prime1030809,The Wall1030810,Danna's Chance1030811,Nopileos' Memorial1030812,Hatikvah's Faith1030813,Open Exchange1030814,Endless Mines1030815,Freedom's Reach1030818,Red Light1030819,Home of Light1030820,President's End1030821,Lasting Vengeance1030903,Blue Storm1030904,Althes1030905,Aldrin1030906,Wastelands1030907,Midnight Star1030908,Belt of Aguilar1030909,Nyana's Hideout1030910,Brennan's Triumph1030911,Rhy's Crusade1030912,Rhy's Desire1030913,Family Pride1030914,Family Kao1030915,Kao's Wolf1030918,Cloudbase South West1030919,Ore Belt1030920,Cloudbase South East1031002,Terran A1031003,Terran B1031004,Megnir1031005,De Vries1031006,Interworlds1031007,Desert Moon1031008,Frozen Ocean1031009,Kingdom End1031010,Xenon Sector 5961031011,Patriarch's Victory1031012,Family Rhy1031013,Rhy's Clouds1031014,Tharka's Sun1031015,Vantera1031016,Aurora1031018,Vestibule of Creation1031019,Green Market1031020,Open Space1031021,Queen's City1031102,Terran D1031103,Terran C1031104,Xenon Sector 1381031105,Unexpected Gift1031106,Mists of Elysium1031107,Great Reef1031108,Reservoir of Transquility1031109,Rolk's Drift1031110,Xenon Sector 5971031111,Lau's Gamble1031112,Family Lau1031113,Grand Cathedral1031114,Xenon Sector 3431031115,Asgard1031116,Hive Remnant1031117,Emperor's Wisdom1031118,Duke's Vision1031119,Clarity's End1031120,Patriarch's Retreat1031202,Queen's Retribution1031203,Hollow Infinity1031204,Desecrated Skies1031205,Gunne's Crusade1031206,Lost River1031207,Two Grand1031208,Nividium Fortune1031209,V Alpha1031210,Home of Opportunity1031211,The Guild1031212,Xaar's Greatness1031213,Nova Paranidia1031214,Mines of Third Duke1031215,Old Realm1031216,Ancient Kingdom1031217,First Colony1031218,Preacher's Void1031219,Third Redemption1031220,Cho's Defeat1031221,Whi's Arrow1031222,Rhonkar's Defeat1031301,Faded Dreams1031302,Queen's Harbour1031303,Menelaus' Oasis1031304,V Gamma1031305,LooManckStrat's Legacy1031306,Creditania1031307,Profit Center Alpha1031308,PTNI Headquarters1031309,Savage Spur1031310,Scale Plate Green1031311,Family Zyarth1031312,Zyarth's Dominion1031313,Zyarth's Stand1031314,Unknown Sector Delta1031316,House of Yamato1031317,Duke's Domain1031318,Empire's Edge1031319,Perdition's End1031320,Patriarch's Keep1031321,Family Whi1031322,Whi's Dragon1031401,Veil of Delusion1031402,Z Alpha1031403,Distant Clouds1031404,Forgotten Realm1031405,Mi Ton's Refuge1031406,New Income1031407,Unlimited Stock1031408,Free Market1031409,Ocracoke's Storm1031410,Atlantia1031411,Family Tkr1031412,Tkr's Deprivation1031413,Ghinn's Escape1031416,New Sky1031417,Aladna Hill1031418,Akeela's Beacon1031419,Gaian Star1031420,Mercenaries' Rift1031421,Maelstrom1031503,Acquisition Repository1031504,Thuruk's Pride1031505,Moo-Kye's Revenge1031506,Ceo's Doubt1031507,V Beta1031508,Teladi Loss1031509,Senator's Badlands1031510,The Giant1031511,Dreamer's Victory1031512,Uranus 21031513,Xenon Sector 71031514,Xenon Sector 91031516,Distant Coast1031517,Light of Heart1031518,Legend's Home1031519,Xenon Sector 1371031520,Lost Order1031521,Twisted Skies1031603,Spaceweed Grove1031604,Rhonkar's Fire1031605,Rhonkar's Mistake1031606,Queen's Space1031607,V Delta1031608,Hunter's Island1031609,Weaver's Tempest1031610,Odin's Moon1031611,Hidden Refuge1031612,Depths of Silence1031613,Dark Waters1031614,Lucky Planets1031616,Albion1031617,Montalaar1031618,Duke's Rebels1031619,Trinity Sanctum1031620,Preacher's Refuge1031621,Spires of Elusion1031703,Unknown Sector Epsilon1031704,Family Rhonkar1031705,Rhonkar's Might1031706,Barren Shores1031707,Menelaus' Frontier1031708,Unknown Sector Alpha1031709,Circle of Labour1031710,Bad Debt1031711,The Vault1031712,Unknown Sector Nu1031713,Shining Currents1031714,Rolk's Fate1031715,Void of Opportunity1031716,Getsu Fune1031717,Leap of Faith1031718,Mining Core1031719,Bright Profit1031720,Priest Refuge1031721,Magrathea1031804,Thuruk's Beard1031805,Xenon Sector 5341031807,Horizon1031808,Federation Core1031809,Omicron Lyrae1031810,Ministry of Finance1031811,Shareholder's Fortune1031812,Mines of Fortune1031813,Xenon Sector 5981031814,Atreus' Clouds1031815,Saturn 21031816,Saturn 31031817,Blue Profit1031818,Sanctity of Corruption1031819,Company Strength1031820,Cardinal's Domain1031821,Sacred Relic1031907,Frontier Gate1031908,Black Hole Sun1031909,Treasure Chest1031910,Homily of Perpetuity1031911,Trader's Peace1031912,Fields of Nividium1031913,Xenon Sector 6271031914,Far Point1031915,Jupiter 21031916,Jupiter 31031917,Ceo's Sprite1031918,Eighteen Billion1031919,Y Beta1031920,Friar's Retreat1031921,Pontifex' Seclusion1032008,Nathan's Voyage1032012,Xenon Sector 6951032013,Xenon Core 0231032014,Argon Sector M1480,Tyrianis Prime1032015,Titan0,Alliance's Edge1032016,Uranus 31032017,Company Pride1032018,Ianamus Zura1032019,Y Alpha0,Unknown Sector
