+++ schemas/readtext_names.csv
@@ -1,0 +1,1 @@
+id,text0,Flaming Dawn0,Evaporated Skies0,Purple Fog0,The Twins0,Dual Eclipse0,Three Moons0,City in Space0,Thor's Might0,Accretion Disc0,Rhonkar's Clouds0,Guiding Star0,Prophecy Fulfilled1020101,Kho's Vengeance1020102,Brennan's Memorial1020117,Duke's Retreat1020118,Fire of Redemption1020121,"{7,1020000} {12,123}"1020122,"{17,10821} {12,110}"1020201,Arcturus Prime1020202,Eye of Colossus1020217,Consecrated Fire1020218,Heaven's Assertion1020222,"{17,10531} {12,109}"1020223,"{17,10431} {12,108}"1020301,Starkiller's Clan1020302,Pilgrim's Relay1020303,Kuiper Belt1020304,Pluto1020305,Neptune1020306,Uranus1020307,Jupiter1020308,Saturn1020312,Scion Fire1020313,Spring of Belief1020314,Perpetual Sin1020315,Harmony1020317,Grand Exchange1020318,Tears of Greed1020319,Elena's Fortune1020320,"{17,10131} {12,100}"1020322,"{17,10231} {12,106}"1020323,"{17,10321} {12,107}"1020401,Starkiller's Hideout1020402,Family Zein1020403,Family Ryk1020404,Oort Cloud1020405,The Moon1020406,Earth1020407,Mars1020408,Family Njy1020409,Njy's Deception1020412,Emperor Mines1020413,Paranid Prime1020414,Priest Rings1020415,Elysium of Light1020416,Unknown Sector Gamma1020417,Merchant Haven1020418,Xenon Sector 3471020419,Priest's Pity1020420,"{17,10261} {12,101}"1020421,"{17,10331} {12,102}"1020422,"{17,10191} {12,105}"1020501,Smuggler's Path1020502,Y Gamma1020503,Rhonkar's Trial1020504,Family Dispute1020505,Mercury1020506,Venus1020507,Asteroid Belt1020508,Thyn's Abyss1020509,Patriarch's Conclusion1020510,Thyn's Excavation1020511,Unseen Domain1020512,Emperor's Pride1020513,Unholy Descent1020514,Ceo's Buckzoid1020515,Teladi Gain1020516,Third Fortune1020517,Lost Profit1020518,Unknown Sector Beta1020519,Emperor's Ridge1020520,Imperial Frontier1020521,"{17,10451} {12,103}"1020522,"{17,10501} {12,104}"1020602,Wretched Skies1020603,The Shallows1020604,Overmasters1020605,Icarus1020606,Heretic's End1020607,Farnham's Legend1020608,Bala Gi's Joy1020609,Olmancketslat's Treaty1020610,Contorted Dominion1020611,Xenon Sector 4721020612,Nomad's Gate1020613,Unknown Sector Kappa1020614,Profit Share1020615,Seizewell1020617,Unholy Gift1020618,Duke's Citadel1020619,Pontifex' Realm1020620,Emperor's Glory1020703,Light Water1020704,Menelaus' Paradise1020705,Rolk's Legacy1020706,Great Trench1020707,Cloudbase North West1020708,Herron's Nebula1020709,The Hole1020710,Split Fire1020711,Sanctuary of Darkness1020712,Xenon Sector 1011020713,Profit Guard1020714,Spaceweed Drift1020715,Greater Profit1020718,Three Worlds1020719,Power Circle1020720,Antigone Memorial1020803,Shore of Infinity1020804,Bluish Snout1020805,Hila's Joy1020806,Ocean of Fantasy1020807,Ringo Moon1020808,Argon Prime1020809,The Wall1020810,Danna's Chance1020811,Nopileos' Memorial1020812,Hatikvah's Faith1020813,Open Exchange1020814,Endless Mines1020815,Freedom's Reach1020818,Red Light1020819,Home of Light1020820,President's End1020821,Lasting Vengeance1020903,Blue Storm1020904,Althes1020905,Aldrin1020906,Wastelands1020907,Midnight Star1020908,Belt of Aguilar1020909,Nyana's Hideout1020910,Brennan's Triumph1020911,Rhy's Crusade1020912,Rhy's Desire1020913,Family Pride1020914,Family Kao1020915,Kao's Wolf1020918,Cloudbase South West1020919,Ore Belt1020920,Cloudbase South East1021002,Terran A1021003,Terran B1021004,Megnir1021005,De Vries1021006,Interworlds1021007,Desert Moon1021008,Frozen Ocean1021009,Kingdom End1021010,Xenon Sector 5961021011,Patriarch's Victory1021012,Family Rhy1021013,Rhy's Clouds1021014,Tharka's Sun1021015,Vantera1021016,Aurora1021018,Vestibule of Creation1021019,Green Market1021020,Open Space1021021,Queen's City1021102,Terran D1021103,Terran C1021104,Xenon Sector 1381021105,Unexpected Gift1021106,Mists of Elysium1021107,Great Reef1021108,Reservoir of Transquility1021109,Rolk's Drift1021110,Xenon Sector 5971021111,Lau's Gamble1021112,Family Lau1021113,Grand Cathedral1021114,Xenon Sector 3431021115,Asgard1021116,Hive Remnant1021117,Emperor's Wisdom1021118,Duke's Vision1021119,Clarity's End1021120,Patriarch's Retreat1021202,Queen's Retribution1021203,Hollow Infinity1021204,Desecrated Skies1021205,Gunne's Crusade1021206,Lost River1021207,Two Grand1021208,Nividium Fortune1021209,V Alpha1021210,Home of Opportunity1021211,The Guild1021212,Xaar's Greatness1021213,Nova Paranidia1021214,Mines of Third Duke1021215,Old Realm1021216,Ancient Kingdom1021217,First Colony1021218,Preacher's Void1021219,Third Redemption1021220,Cho's Defeat1021221,Whi's Arrow1021222,Rhonkar's Defeat1021301,Faded Dreams1021302,Queen's Harbour1021303,Menelaus' Oasis1021304,V Gamma1021305,LooManckStrat's Legacy1021306,Creditania1021307,Profit Center Alpha1021308,PTNI Headquarters1021309,Savage Spur1021310,Scale Plate Green1021311,Family Zyarth1021312,Zyarth's Dominion1021313,Zyarth's Stand1021314,Unknown Sector Delta1021316,House of Yamato1021317,Duke's Domain1021318,Empire's Edge1021319,Perdition's End1021320,Patriarch's Keep1021321,Family Whi1021322,Whi's Dragon1021401,Veil of Delusion1021402,Z Alpha1021403,Distant Clouds1021404,Forgotten Realm1021405,Mi Ton's Refuge1021406,New Income1021407,Unlimited Stock1021408,Free Market1021409,Ocracoke's Storm1021410,Atlantia1021411,Family Tkr1021412,Tkr's Deprivation1021413,Ghinn's Escape1021416,New Sky1021417,Aladna Hill1021418,Akeela's Beacon1021419,Gaian Star1021420,Mercenaries' Rift1021421,Maelstrom1021503,Acquisition Repository1021504,Thuruk's Pride1021505,Moo-Kye's Revenge1021506,Ceo's Doubt1021507,V Beta1021508,Teladi Loss1021509,Senator's Badlands1021510,The Giant1021511,Dreamer's Victory1021512,Uranus 21021513,Xenon Sector 71021514,Xenon Sector 91021516,Distant Coast1021517,Light of Heart1021518,Legend's Home1021519,Xenon Sector 1371021520,Lost Order1021521,Twisted Skies1021603,Spaceweed Grove1021604,Rhonkar's Fire1021605,Rhonkar's Mistake1021606,Queen's Space1021607,V Delta1021608,Hunter's Island1021609,Weaver's Tempest1021610,Odin's Moon1021611,Hidden Refuge1021612,Depths of Silence1021613,Dark Waters1021614,Lucky Planets1021616,Albion1021617,Montalaar1021618,Duke's Rebels1021619,Trinity Sanctum1021620,Preacher's Refuge1021621,Spires of Elusion1021703,Unknown Sector Epsilon1021704,Family Rhonkar1021705,Rhonkar's Might1021706,Barren Shores1021707,Menelaus' Frontier1021708,Unknown Sector Alpha1021709,Circle of Labour1021710,Bad Debt1021711,The Vault1021712,Unknown Sector Nu1021713,Shining Currents1021714,Rolk's Fate1021715,Void of Opportunity1021716,Getsu Fune1021717,Leap of Faith1021718,Mining Core1021719,Bright Profit1021720,Priest Refuge1021721,Magrathea1021804,Thuruk's Beard1021805,Xenon Sector 5341021807,Horizon1021808,Federation Core1021809,Omicron Lyrae1021810,Ministry of Finance1021811,Shareholder's Fortune1021812,Mines of Fortune1021813,Xenon Sector 5981021814,Atreus' Clouds1021815,Saturn 21021816,Saturn 31021817,Blue Profit1021818,Sanctity of Corruption1021819,Company Strength1021820,Cardinal's Domain1021821,Sacred Relic1021907,Frontier Gate1021908,Black Hole Sun1021909,Treasure Chest1021910,Homily of Perpetuity1021911,Trader's Peace1021912,Fields of Nividium1021913,Xenon Sector 6271021914,Far Point1021915,Jupiter 21021916,Jupiter 31021917,Ceo's Sprite1021918,Eighteen Billion1021919,Y Beta1021920,Friar's Retreat1021921,Pontifex' Seclusion1022008,Nathan's Voyage1022012,Xenon Sector 6951022013,Xenon Core 0231022014,Argon Sector M1480,Tyrianis Prime1022015,Titan0,Alliance's Edge1022016,Uranus 31022017,Company Pride1022018,Ianamus Zura1022019,Y Alpha1020000,Unknown Sector
